upload + w3c + url etc

+++ views/dashboard/users/utilisateurs.php
@@ -6,23 +6,23 @@
         </a>
         <h1 class="text-center mb-4 ">Utilisateurs</h1>
         <!-- <p class="text-center info-color fs-3 fw-bold"> -->
-            <!-- message informatif, variable $delete / $archive définie dans le controlleur category -->
-            <!-- cible l'action dans le paramètre d'url -->
-            <!-- <?php
-            if ($delete === '1') {
-                echo 'Catégorie supprimée avec succès';
-            } else if ($delete === '0') {
-                echo 'Erreur pendant la suppression de la catégorie';
-            } else if ($archive === '1') {
-                echo 'Élément archivé avec succès';
-            } else if ($archive === '0') {
-                echo 'Erreur pendant l\'archivage de l\'élément';
-            } else if ($unarchive === '1') {
-                echo 'Élement désarchivé avec succès';
-            } else if ($unarchive === '0') {
-                echo 'Erreur pendant le désarchivage de l\'élément';
-            }
-            ?> -->
+        <!-- message informatif, variable $delete / $archive définie dans le controlleur category -->
+        <!-- cible l'action dans le paramètre d'url -->
+        <!-- <?php
+                if ($delete === '1') {
+                    echo 'Catégorie supprimée avec succès';
+                } else if ($delete === '0') {
+                    echo 'Erreur pendant la suppression de la catégorie';
+                } else if ($archive === '1') {
+                    echo 'Élément archivé avec succès';
+                } else if ($archive === '0') {
+                    echo 'Erreur pendant l\'archivage de l\'élément';
+                } else if ($unarchive === '1') {
+                    echo 'Élement désarchivé avec succès';
+                } else if ($unarchive === '0') {
+                    echo 'Erreur pendant le désarchivage de l\'élément';
+                }
+                ?> -->
         <!-- </p> -->
         <div class="row">
             <div class="col d-flex justify-content-center h-50">
@@ -102,14 +102,31 @@
                                             </svg>
                                         </button>
                                     </a>
-                                    <a href="/controllers/dashboard/users/delete_utilisateurs_controller.php?action=delete&id_users=<?= $userArchived->id_users ?>">
-                                        <button class="btn btn-transparent" title="Supprimer l'élément">
-                                            <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" fill="currentColor" class="bi bi-x-circle" viewBox="0 0 16 16">
-                                                <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
-                                                <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z" />
-                                            </svg>
-                                        </button>
-                                    </a>
+                                    <!-- Modal de confirmation de suppression -->
+                                    <div class="modal" id="confirmationModal" tabindex="-1" aria-labelledby="confirmationModalLabel" aria-hidden="true">
+                                        <div class="modal-dialog">
+                                            <div class="modal-content">
+                                                <div class="modal-header">
+                                                    <h5 class="modal-title" id="confirmationModalLabel">Confirmation de suppression</h5>
+                                                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                                </div>
+                                                <div class="modal-body">
+                                                    Êtes-vous sûr de vouloir supprimer <span class="fw-bold"><?= $userArchived->username ?></span> ?
+                                                </div>
+                                                <div class="modal-footer">
+                                                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Annuler</button>
+                                                    <a id="confirmDeleteButton" href="/controllers/dashboard/users/delete_utilisateurs_controller.php?action=delete&id_users=<?= $userArchived->id_users ?>" class="btn btn-danger">Confirmer la suppression</a>
+                                                </div>
+                                            </div>
+                                        </div>
+                                    </div>
+                                    <!-- Bouton pour ouvrir la modal -->
+                                    <button type="button" class="btn btn-transparent delete-button" title="Supprimer l'élément" data-bs-toggle="modal" data-bs-target="#confirmationModal">
+                                        <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" fill="currentColor" class="bi bi-x-circle" viewBox="0 0 16 16">
+                                            <path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z" />
+                                            <path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z" />
+                                        </svg>
+                                    </button>
                                 </td>
                             </tr>
                         <?php } ?>
